+++ src/exchanges/gate/gateSpecs.ts
@@ -45,10 +45,6 @@ export const gateBaseSpecs: IAlunaExchangeSchema = {
     offersOrderEditing: false,
     offersPositionId: false,
   },
-  modes: {
-    balance: AlunaFeaturesModeEnum.READ,
-    order: AlunaFeaturesModeEnum.WRITE,
-  },
   accounts: [
     {
       type: AlunaAccountEnum.SPOT,