+++ src/exchanges/poloniex/modules/authed/order/place.spec.ts
@@ -20,6 +20,7 @@ import { PoloniexOrderTimeInForceEnum } from '../../../enums/PoloniexOrderTimeIn
 import { PoloniexAuthed } from '../../../PoloniexAuthed'
 import { PoloniexHttp } from '../../../PoloniexHttp'
 import { getPoloniexEndpoints } from '../../../poloniexSpecs'
+import { IPoloniexOrderPlaceResponseSchema } from '../../../schemas/IPoloniexOrderSchema'
 import { POLONIEX_RAW_ORDERS } from '../../../test/fixtures/poloniexOrders'
 import * as getMod from './get'
 
@@ -114,32 +115,22 @@ describe(__filename, () => {
     'should throw error for insufficient funds when placing new poloniex order',
     async () => {
 
-      // preparing data
-      // const mockedRawOrder = POLONIEX_RAW_ORDERS[0]
-
-      const side = AlunaOrderSideEnum.BUY
-      const type = AlunaOrderTypesEnum.MARKET
-
-      const expectedMessage = 'Account has insufficient balance '
-        .concat('for requested action.')
-      const expectedCode = AlunaBalanceErrorCodes.INSUFFICIENT_BALANCE
-
-      const alunaError = new AlunaError({
-        message: 'Not enough',
-        code: AlunaOrderErrorCodes.PLACE_FAILED,
-        httpStatusCode: 401,
-        metadata: {
-          code: 'INSUFFICIENT_FUNDS',
-        },
-      })
-
       // mocking
+      const placeResponse: IPoloniexOrderPlaceResponseSchema = {
+        clientOrderId: 'clientOrderId',
+        fee: 'fee',
+        currencyPair: 'BTC_USDC',
+        orderNumber: '666',
+        resultingTrades: [],
+        error: 'Not enough USDC.',
+      }
+
       const {
         publicRequest,
         authedRequest,
       } = mockHttp({ classPrototype: PoloniexHttp.prototype })
 
-      authedRequest.returns(Promise.reject(alunaError))
+      authedRequest.returns(Promise.resolve(placeResponse))
 
       mockValidateParams()
 
@@ -153,17 +144,69 @@ describe(__filename, () => {
         symbolPair: '',
         account: AlunaAccountEnum.SPOT,
         amount: 0.01,
-        side,
-        type,
+        side: AlunaOrderSideEnum.BUY,
+        type: AlunaOrderTypesEnum.MARKET,
         rate: 0,
       }))
 
 
       // validating
 
-      expect(error instanceof AlunaError).to.be.ok
-      expect(error?.code).to.be.eq(expectedCode)
-      expect(error?.message).to.be.eq(expectedMessage)
+      expect(error!.code).to.be.eq(AlunaBalanceErrorCodes.INSUFFICIENT_BALANCE)
+      expect(error!.message).to.be.eq(placeResponse.error)
+      expect(error!.metadata).to.be.eq(placeResponse)
+
+      expect(authedRequest.callCount).to.be.eq(1)
+
+      expect(publicRequest.callCount).to.be.eq(0)
+
+    },
+  )
+
+  it(
+    'should throw with proper error code when order placement fails somehow',
+    async () => {
+
+      // mocking
+      const placeResponse: IPoloniexOrderPlaceResponseSchema = {
+        clientOrderId: 'clientOrderId',
+        fee: 'fee',
+        currencyPair: 'BTC_USDC',
+        orderNumber: '666',
+        resultingTrades: [],
+        error: 'Exchange offline',
+      }
+
+      const {
+        publicRequest,
+        authedRequest,
+      } = mockHttp({ classPrototype: PoloniexHttp.prototype })
+
+      authedRequest.returns(Promise.resolve(placeResponse))
+
+      mockValidateParams()
+
+      mockEnsureOrderIsSupported()
+
+
+      // executing
+      const exchange = new PoloniexAuthed({ credentials })
+
+      const { error } = await executeAndCatch(() => exchange.order.place({
+        symbolPair: '',
+        account: AlunaAccountEnum.SPOT,
+        amount: 0.01,
+        side: AlunaOrderSideEnum.BUY,
+        type: AlunaOrderTypesEnum.MARKET,
+        rate: 0,
+      }))
+
+
+      // validating
+
+      expect(error!.code).to.be.eq(AlunaOrderErrorCodes.PLACE_FAILED)
+      expect(error!.message).to.be.eq(placeResponse.error)
+      expect(error!.metadata).to.be.eq(placeResponse)
 
       expect(authedRequest.callCount).to.be.eq(1)