+++ src/exchanges/valr/valrSpecs.ts
@@ -77,10 +77,6 @@ export const valrBaseSpecs: IAlunaExchangeSchema = {
     offersOrderEditing: false,
     offersPositionId: false,
   },
-  modes: {
-    balance: AlunaFeaturesModeEnum.READ,
-    order: AlunaFeaturesModeEnum.WRITE,
-  },
   accounts: [
     {
       type: AlunaAccountEnum.SPOT,