+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+# 1.0.2 (May 19, 2022)
+
+ - Filtering zeroed balances
+ - Fixing insufficient-balance error for Valr/Poloniex
+
+# 1.0.1 (May 18, 2022)
+
+ - Adjustments in `IAlunaExchangeSchema` interface
+   - Removes obsolete [*and broken*] modes property
+   - Making property features required (*as it should*)
+
 # 1.0.0 (May 17, 2022)
 
  - Initial public release, includes: